moved functions around to better places
