+++ src/app/features/completion/utils/playing.utils.spec.ts
@@ -1,105 +1,9 @@
 import { Dictionary, HistoryGrouping, HistoryListItem } from '../../../shared/models';
 import { MarkCompleteItem, ProgressItem } from '../models';
 
-import {
-    filterHistoryItemsFrom, getEndItem, getHistoryListItemMap, getPlayingDisplayData,
-    getPlayingItem, HistoryListItemMap
-} from './playing.utils';
+import { getEndItem, getPlayingDisplayData, getPlayingItem } from './playing.utils';
 
 describe('Playing Utils', () => {
-  describe('getHistoryListItemMap', () => {
-    it('Should return empty when there are no groupings', () => {
-      const result = getHistoryListItemMap([]);
-      expect(result).toEqual(new Map<string, HistoryListItem>());
-    });
-
-    it('Should return correct map when there are groupings', () => {
-      const expected: HistoryListItemMap = new Map<string, HistoryListItem>();
-      expected.set('3', {
-        id: '3',
-        game: testGame,
-        platform: 'Platform 1',
-        startTime: 5000,
-        endTime: 6000,
-        dateRange: [new Date(5000), new Date(6000)]
-      });
-      expected.set('2', {
-        id: '2',
-        game: testGame,
-        platform: 'Platform 1',
-        startTime: 3000,
-        endTime: 4000,
-        dateRange: [new Date(3000), new Date(4000)]
-      });
-      expected.set('1', {
-        id: '1',
-        game: testGame,
-        platform: 'Platform 2',
-        startTime: 1000,
-        endTime: 2000,
-        dateRange: [new Date(1000), new Date(2000)]
-      });
-      const result = getHistoryListItemMap(testGroupings);
-      expect(result).toEqual(expected);
-    });
-  });
-
-  describe('filterHistoryItemsFrom', () => {
-    it('Should return empty if there are no groupings', () => {
-      const result = filterHistoryItemsFrom({
-        key: testGame,
-        historyItems: [],
-        totalTime: 0
-      }, {
-          id: 'some id',
-          game: testGame,
-          platform: 'Platform 1',
-          startTime: 6000,
-          endTime: 9000,
-          dateRange: [new Date(6000), new Date(9000)]
-        });
-      expect(result.length).toBe(0);
-    });
-
-    it('Should return empty if there are groupings but no match', () => {
-      const result = filterHistoryItemsFrom(testGroupings[0], {
-        id: 'some id',
-        game: testGame,
-        platform: 'Platform 1',
-        startTime: 6000,
-        endTime: 9000,
-        dateRange: [new Date(6000), new Date(9000)]
-      });
-      expect(result.length).toBe(0);
-    });
-
-    it('Should return correctly filtered items when there is a match', () => {
-      const result = filterHistoryItemsFrom(testGroupings[0], {
-        id: 'some id',
-        game: testGame,
-        platform: 'Platform 1',
-        startTime: 3000,
-        endTime: 4000,
-        dateRange: [new Date(3000), new Date(4000)]
-      });
-      expect(result).toEqual([{
-        id: '3',
-        game: testGame,
-        platform: 'Platform 1',
-        startTime: 5000,
-        endTime: 6000,
-        dateRange: [new Date(5000), new Date(6000)]
-      }, {
-        id: '2',
-        game: testGame,
-        platform: 'Platform 1',
-        startTime: 3000,
-        endTime: 4000,
-        dateRange: [new Date(3000), new Date(4000)]
-      }]);
-    });
-  });
-
   describe('getPlayingItem', () => {
     it('Should return correct data', () => {
       const result = getPlayingItem(testGroupings[0].historyItems, {
@@ -240,4 +144,3 @@ const testGroupings: HistoryGrouping[] = [{
   }],
   totalTime: 3
 }];
-
