+++ src/app/shared/utils/history-filter.utils.spec.ts
@@ -1,7 +1,7 @@
 import { HistoryGrouping, HistoryListItem } from '../models';
 
 import {
-    filterEndTimes, filterPlatforms, filterStartTimes, getUniqueFrom
+    filterEndTimes, filterHistoryItemsFrom, filterPlatforms, filterStartTimes, getUniqueFrom
 } from './history-filter.utils';
 
 describe('History Filter Utils', () => {
@@ -192,6 +192,62 @@ describe('History Filter Utils', () => {
       expect(result).toEqual([5000]);
     });
   });
+
+  describe('filterHistoryItemsFrom', () => {
+    it('Should return empty if there are no groupings', () => {
+      const result = filterHistoryItemsFrom({
+        key: testGame,
+        historyItems: [],
+        totalTime: 0
+      }, {
+          id: 'some id',
+          game: testGame,
+          platform: 'Platform 1',
+          startTime: 6000,
+          endTime: 9000,
+          dateRange: [new Date(6000), new Date(9000)]
+        });
+      expect(result.length).toBe(0);
+    });
+
+    it('Should return empty if there are groupings but no match', () => {
+      const result = filterHistoryItemsFrom(testGroupings[0], {
+        id: 'some id',
+        game: testGame,
+        platform: 'Platform 1',
+        startTime: 6000,
+        endTime: 9000,
+        dateRange: [new Date(6000), new Date(9000)]
+      });
+      expect(result.length).toBe(0);
+    });
+
+    it('Should return correctly filtered items when there is a match', () => {
+      const result = filterHistoryItemsFrom(testGroupings[0], {
+        id: 'some id',
+        game: testGame,
+        platform: 'Platform 1',
+        startTime: 3000,
+        endTime: 4000,
+        dateRange: [new Date(3000), new Date(4000)]
+      });
+      expect(result).toEqual([{
+        id: '3',
+        game: testGame,
+        platform: 'Platform 1',
+        startTime: 5000,
+        endTime: 6000,
+        dateRange: [new Date(5000), new Date(6000)]
+      }, {
+        id: '2',
+        game: testGame,
+        platform: 'Platform 1',
+        startTime: 3000,
+        endTime: 4000,
+        dateRange: [new Date(3000), new Date(4000)]
+      }]);
+    });
+  });
 });
 
 const testGame = 'Game 1';
@@ -214,3 +270,30 @@ const getHistoryGrouping = (key: string, totalTime: number): HistoryGrouping => 
     historyItems: []
   };
 };
+
+const testGroupings: HistoryGrouping[] = [{
+  key: testGame,
+  historyItems: [{
+    id: '3',
+    game: testGame,
+    platform: 'Platform 1',
+    startTime: 5000,
+    endTime: 6000,
+    dateRange: [new Date(5000), new Date(6000)]
+  }, {
+    id: '2',
+    game: testGame,
+    platform: 'Platform 1',
+    startTime: 3000,
+    endTime: 4000,
+    dateRange: [new Date(3000), new Date(4000)]
+  }, {
+    id: '1',
+    game: testGame,
+    platform: 'Platform 2',
+    startTime: 1000,
+    endTime: 2000,
+    dateRange: [new Date(1000), new Date(2000)]
+  }],
+  totalTime: 3
+}];
